docs(IconButton): update public documentation + prop API
